+++ yandex/mdb_mongodb_structures.go
@@ -16,8 +16,10 @@ import (
 )
 
 var supportedVersions = map[string]bool{
+	"6.0-enterprise": true,
 	"5.0-enterprise": true,
 	"4.4-enterprise": true,
+	"6.0":            true,
 	"5.0":            true,
 	"4.4":            true,
 	"4.2":            true,
@@ -35,6 +37,667 @@ type MongodbSpecHelper struct {
 
 func GetMongodbSpecHelper(version string) *MongodbSpecHelper {
 	switch version {
+	case "6.0-enterprise":
+		{
+			return &MongodbSpecHelper{
+				FlattenResources: func(c *mongodb.ClusterConfig, d *schema.ResourceData) (map[string]interface{}, error) {
+					spec := c.Mongodb.(*mongodb.ClusterConfig_Mongodb_6_0Enterprise).Mongodb_6_0Enterprise
+					resources := map[string]interface{}{}
+					if _, ok := d.GetOk("resources"); ok {
+						if spec.Mongod != nil {
+							resources["resources"] = flattenMongoDBResources(spec.Mongod.Resources)
+							return resources, nil
+						}
+						if spec.Mongos != nil {
+							resources["resources"] = flattenMongoDBResources(spec.Mongos.Resources)
+							return resources, nil
+						}
+						if spec.Mongocfg != nil {
+							resources["resources"] = flattenMongoDBResources(spec.Mongocfg.Resources)
+							return resources, nil
+						}
+						if spec.Mongoinfra != nil {
+							resources["resources"] = flattenMongoDBResources(spec.Mongoinfra.Resources)
+							return resources, nil
+						}
+					} else {
+						if spec.Mongod != nil {
+							resources["resources_mongod"] = flattenMongoDBResources(spec.Mongod.Resources)
+						}
+						if spec.Mongos != nil {
+							resources["resources_mongos"] = flattenMongoDBResources(spec.Mongos.Resources)
+						}
+						if spec.Mongocfg != nil {
+							resources["resources_mongocfg"] = flattenMongoDBResources(spec.Mongocfg.Resources)
+						}
+						if spec.Mongoinfra != nil {
+							resources["resources_mongoinfra"] = flattenMongoDBResources(spec.Mongoinfra.Resources)
+						}
+					}
+					if len(resources) == 0 {
+						return nil, fmt.Errorf("Non empty service not found in mongo spec")
+					}
+					return resources, nil
+				},
+
+				FlattenMongod: func(c *mongodb.ClusterConfig, d *schema.ResourceData) ([]map[string]interface{}, error) {
+					mongod := c.Mongodb.(*mongodb.ClusterConfig_Mongodb_6_0Enterprise).Mongodb_6_0Enterprise.Mongod
+					if mongod != nil {
+						user_config := mongod.GetConfig().GetUserConfig()
+						default_config := mongod.GetConfig().GetDefaultConfig()
+
+						result := map[string]interface{}{}
+
+						if security := user_config.GetSecurity(); security != nil {
+							flattenSecurity := map[string]interface{}{}
+							if enableEncription := security.GetEnableEncryption(); enableEncription != nil {
+								flattenSecurity["enable_encryption"] = enableEncription.GetValue()
+							}
+							if kmip := security.GetKmip(); kmip != nil {
+								flattenKmip := map[string]interface{}{}
+								flattenKmip["server_name"] = kmip.GetServerName()
+								flattenKmip["port"] = int(kmip.GetPort().GetValue())
+								flattenKmip["server_ca"] = kmip.GetServerCa()
+								flattenKmip["client_certificate"] = d.Get("cluster_config.0.mongod.0.security.0.kmip.0.client_certificate")
+								flattenKmip["key_identifier"] = kmip.GetKeyIdentifier()
+
+								flattenSecurity["kmip"] = []map[string]interface{}{flattenKmip}
+							}
+							result["security"] = []map[string]interface{}{flattenSecurity}
+						}
+
+						if audit_log := user_config.GetAuditLog(); audit_log != nil {
+							audit_log_data := map[string]interface{}{}
+							if audit_log.GetFilter() != default_config.GetAuditLog().GetFilter() {
+								audit_log_data["filter"] = audit_log.GetFilter()
+							}
+							if audit_log.GetRuntimeConfiguration() != nil {
+								audit_log_data["runtime_configuration"] = audit_log.GetRuntimeConfiguration().GetValue()
+							}
+							result["audit_log"] = []map[string]interface{}{audit_log_data}
+						}
+						if set_parameter := user_config.GetSetParameter(); set_parameter != nil {
+							set_parameter_data := map[string]interface{}{}
+							if set_parameter.GetAuditAuthorizationSuccess() != nil {
+								set_parameter_data["audit_authorization_success"] = set_parameter.GetAuditAuthorizationSuccess().GetValue()
+							}
+							result["set_parameter"] = []map[string]interface{}{set_parameter_data}
+						}
+
+						if net := user_config.GetNet(); net != nil {
+							flattenNet := map[string]interface{}{}
+							if maxIncomingConnections := net.GetMaxIncomingConnections(); maxIncomingConnections != nil {
+								flattenNet["max_incoming_connections"] = maxIncomingConnections.GetValue()
+							}
+							result["net"] = []map[string]interface{}{flattenNet}
+						}
+
+						if storage := user_config.GetStorage(); storage != nil {
+							flattenStorage := map[string]interface{}{}
+							if wiredTiger := storage.GetWiredTiger(); wiredTiger != nil {
+								flattenWiredTiger := map[string]interface{}{}
+								if engineConfig := wiredTiger.GetEngineConfig(); engineConfig != nil {
+									if cacheSize := engineConfig.GetCacheSizeGb(); cacheSize != nil {
+										flattenWiredTiger["cache_size_gb"] = cacheSize.GetValue()
+									}
+								}
+								if collectionConfig := wiredTiger.GetCollectionConfig(); collectionConfig != nil {
+									if blockCompressor := collectionConfig.GetBlockCompressor(); blockCompressor != 0 {
+										flattenWiredTiger["block_compressor"] = blockCompressor.String()
+									}
+								}
+								flattenStorage["wired_tiger"] = []map[string]interface{}{flattenWiredTiger}
+							}
+
+							if journal := storage.GetJournal(); journal != nil {
+								flattenJournal := map[string]interface{}{}
+								if commitInterval := journal.GetCommitInterval(); commitInterval != nil {
+									flattenJournal["commit_interval"] = commitInterval.GetValue()
+								}
+								flattenStorage["journal"] = []map[string]interface{}{flattenJournal}
+							}
+							result["storage"] = []map[string]interface{}{flattenStorage}
+						}
+
+						if opProfiling := user_config.GetOperationProfiling(); opProfiling != nil {
+							flattenOpProfiling := map[string]interface{}{}
+							if mode := opProfiling.GetMode(); mode != 0 {
+								flattenOpProfiling["mode"] = mode.String()
+							}
+							if opThreshold := opProfiling.GetSlowOpThreshold(); opThreshold != nil {
+								flattenOpProfiling["slow_op_threshold"] = opThreshold.GetValue()
+							}
+							result["operation_profiling"] = []map[string]interface{}{flattenOpProfiling}
+						}
+
+						return []map[string]interface{}{result}, nil
+					}
+					return []map[string]interface{}{}, nil
+				},
+
+				FlattenMongos: func(c *mongodb.ClusterConfig, d *schema.ResourceData) ([]map[string]interface{}, error) {
+					mongos := c.Mongodb.(*mongodb.ClusterConfig_Mongodb_6_0Enterprise).Mongodb_6_0Enterprise.Mongos
+					if mongos != nil {
+						userConfig := mongos.GetConfig().GetUserConfig()
+						result := map[string]interface{}{}
+
+						if net := userConfig.GetNet(); net != nil {
+							flattenNet := map[string]interface{}{}
+							if maxIncomingConnections := net.GetMaxIncomingConnections(); maxIncomingConnections != nil {
+								flattenNet["max_incoming_connections"] = maxIncomingConnections.GetValue()
+							}
+							result["net"] = []map[string]interface{}{flattenNet}
+						}
+						return []map[string]interface{}{result}, nil
+					}
+					return []map[string]interface{}{}, nil
+				},
+
+				FlattenMongocfg: func(c *mongodb.ClusterConfig, d *schema.ResourceData) ([]map[string]interface{}, error) {
+					mongocfg := c.Mongodb.(*mongodb.ClusterConfig_Mongodb_6_0Enterprise).Mongodb_6_0Enterprise.Mongocfg
+					if mongocfg != nil {
+						userConfig := mongocfg.GetConfig().GetUserConfig()
+						result := map[string]interface{}{}
+
+						if net := userConfig.GetNet(); net != nil {
+							flattenNet := map[string]interface{}{}
+							if maxIncomingConnections := net.GetMaxIncomingConnections(); maxIncomingConnections != nil {
+								flattenNet["max_incoming_connections"] = maxIncomingConnections.GetValue()
+							}
+							result["net"] = []map[string]interface{}{flattenNet}
+						}
+
+						if storage := userConfig.GetStorage(); storage != nil {
+							flattenStorage := map[string]interface{}{}
+							if wiredTiger := storage.GetWiredTiger(); wiredTiger != nil {
+								flattenWiredTiger := map[string]interface{}{}
+								if engineConfig := wiredTiger.GetEngineConfig(); engineConfig != nil {
+									if cacheSize := engineConfig.GetCacheSizeGb(); cacheSize != nil {
+										flattenWiredTiger["cache_size_gb"] = cacheSize.GetValue()
+									}
+								}
+								flattenStorage["wired_tiger"] = []map[string]interface{}{flattenWiredTiger}
+							}
+							result["storage"] = []map[string]interface{}{flattenStorage}
+						}
+
+						if opProfiling := userConfig.GetOperationProfiling(); opProfiling != nil {
+							flattenOpProfiling := map[string]interface{}{}
+							if mode := opProfiling.GetMode(); mode != 0 {
+								flattenOpProfiling["mode"] = mode.String()
+							}
+							if opThreshold := opProfiling.GetSlowOpThreshold(); opThreshold != nil {
+								flattenOpProfiling["slow_op_threshold"] = opThreshold.GetValue()
+							}
+							result["operation_profiling"] = []map[string]interface{}{flattenOpProfiling}
+						}
+
+						return []map[string]interface{}{result}, nil
+					}
+					return []map[string]interface{}{}, nil
+				},
+
+				Expand: func(d *schema.ResourceData) mongodb.ConfigSpec_MongodbSpec {
+					configMongod := mongo_config.MongodConfig6_0Enterprise{}
+					configMongos := mongo_config.MongosConfig6_0Enterprise{}
+					configMongoCfg := mongo_config.MongoCfgConfig6_0Enterprise{}
+
+					security := mongo_config.MongodConfig6_0Enterprise_Security{}
+					if enable_encryption := d.Get("cluster_config.0.mongod.0.security.0.enable_encryption"); enable_encryption != nil {
+						security.SetEnableEncryption(&wrappers.BoolValue{Value: enable_encryption.(bool)})
+					}
+					kmip := mongo_config.MongodConfig6_0Enterprise_Security_KMIP{}
+					if server_name := d.Get("cluster_config.0.mongod.0.security.0.kmip.0.server_name"); server_name != nil {
+						kmip.SetServerName(server_name.(string))
+					}
+					if port := d.Get("cluster_config.0.mongod.0.security.0.kmip.0.port"); port != nil {
+						kmip.SetPort(&wrappers.Int64Value{Value: int64(port.(int))})
+					}
+					if server_ca := d.Get("cluster_config.0.mongod.0.security.0.kmip.0.server_ca"); server_ca != nil {
+						kmip.SetServerCa(server_ca.(string))
+					}
+					if client_certificate := d.Get("cluster_config.0.mongod.0.security.0.kmip.0.client_certificate"); client_certificate != nil {
+						kmip.SetClientCertificate(client_certificate.(string))
+					}
+					if key_identifier := d.Get("cluster_config.0.mongod.0.security.0.kmip.0.key_identifier"); key_identifier != nil {
+						kmip.SetKeyIdentifier(key_identifier.(string))
+					}
+					security.SetKmip(&kmip)
+					configMongod.SetSecurity(&security)
+					audit_log := mongo_config.MongodConfig6_0Enterprise_AuditLog{}
+					if filter := d.Get("cluster_config.0.mongod.0.audit_log.0.filter"); filter != nil {
+						audit_log.SetFilter(filter.(string))
+					}
+					// Note: right now runtime_configuration unsupported, so we should comment this statement
+					//if rt := d.Get("cluster_config.0.mongod.0.audit_log.0.runtime_configuration"); rt != nil {
+					//	audit_log.SetRuntimeConfiguration(&wrappers.BoolValue{Value: rt.(bool)})
+					//}
+					configMongod.SetAuditLog(&audit_log)
+
+					set_paramenter := mongo_config.MongodConfig6_0Enterprise_SetParameter{}
+					if success := d.Get("cluster_config.0.mongod.0.set_parameter.0.audit_authorization_success"); success != nil {
+						set_paramenter.SetAuditAuthorizationSuccess(&wrappers.BoolValue{Value: success.(bool)})
+					}
+					configMongod.SetSetParameter(&set_paramenter)
+					if _, ok := d.GetOk("cluster_config.0.mongod.0.net"); ok {
+						netMongod := mongo_config.MongodConfig6_0Enterprise_Network{}
+						if maxConnections, ok := d.GetOk("cluster_config.0.mongod.0.net.0.max_incoming_connections"); ok {
+							netMongod.SetMaxIncomingConnections(&wrappers.Int64Value{Value: int64(maxConnections.(int))})
+						}
+						configMongod.SetNet(&netMongod)
+					}
+					if _, ok := d.GetOk("cluster_config.0.mongod.0.operation_profiling"); ok {
+						opProfilingMongod := mongo_config.MongodConfig6_0Enterprise_OperationProfiling{}
+
+						if mode, ok := d.GetOk("cluster_config.0.mongod.0.operation_profiling.0.mode"); ok {
+							modeInt := mongo_config.MongodConfig6_0_OperationProfiling_Mode_value[strings.ToUpper(mode.(string))]
+							opProfilingMongod.SetMode(mongo_config.MongodConfig6_0Enterprise_OperationProfiling_Mode(modeInt))
+						}
+
+						if opThreshold, ok := d.GetOk("cluster_config.0.mongod.0.operation_profiling.0.slow_op_threshold"); ok {
+							opProfilingMongod.SetSlowOpThreshold(&wrappers.Int64Value{Value: int64(opThreshold.(int))})
+						}
+						configMongod.SetOperationProfiling(&opProfilingMongod)
+					}
+					if _, ok := d.GetOk("cluster_config.0.mongod.0.storage"); ok {
+						engineConfigMongod := mongo_config.MongodConfig6_0Enterprise_Storage_WiredTiger_EngineConfig{}
+						collectionConfigMongod := mongo_config.MongodConfig6_0Enterprise_Storage_WiredTiger_CollectionConfig{}
+						journalMongod := mongo_config.MongodConfig6_0Enterprise_Storage_Journal{}
+						wiredTigerMongod := mongo_config.MongodConfig6_0Enterprise_Storage_WiredTiger{
+							EngineConfig:     &engineConfigMongod,
+							CollectionConfig: &collectionConfigMongod,
+						}
+						storageMongod := mongo_config.MongodConfig6_0Enterprise_Storage{
+							WiredTiger: &wiredTigerMongod,
+							Journal:    &journalMongod,
+						}
+						if cacheSize, ok := d.GetOk("cluster_config.0.mongod.0.storage.0.wired_tiger.0.cache_size_gb"); ok {
+							engineConfigMongod.SetCacheSizeGb(&wrappers.DoubleValue{Value: float64(cacheSize.(float64))})
+						}
+						if blockCompressor, ok := d.GetOk("cluster_config.0.mongod.0.storage.0.wired_tiger.0.block_compressor"); ok {
+							blockCompressorInt := mongo_config.MongodConfig6_0Enterprise_Storage_WiredTiger_CollectionConfig_Compressor_value[strings.ToUpper(blockCompressor.(string))]
+							collectionConfigMongod.SetBlockCompressor(
+								mongo_config.MongodConfig6_0Enterprise_Storage_WiredTiger_CollectionConfig_Compressor(blockCompressorInt),
+							)
+						}
+
+						if commitInterval, ok := d.GetOk("cluster_config.0.mongod.0.storage.0.journal.0.commit_interval"); ok {
+							journalMongod.SetCommitInterval(&wrappers.Int64Value{Value: int64(commitInterval.(int))})
+						}
+						configMongod.SetStorage(&storageMongod)
+					}
+					if _, ok := d.GetOk("cluster_config.0.mongos.0.net"); ok {
+
+						netMongos := mongo_config.MongosConfig6_0Enterprise_Network{}
+						if maxConnections, ok := d.GetOk("cluster_config.0.mongos.0.net.0.max_incoming_connections"); ok {
+							netMongos.SetMaxIncomingConnections(&wrappers.Int64Value{Value: int64(maxConnections.(int))})
+						}
+						configMongos.SetNet(&netMongos)
+					}
+					if _, ok := d.GetOk("cluster_config.0.mongocfg.0.net"); ok {
+						netMongoCfg := mongo_config.MongoCfgConfig6_0Enterprise_Network{}
+						if maxConnections, ok := d.GetOk("cluster_config.0.mongocfg.0.net.0.max_incoming_connections"); ok {
+							netMongoCfg.SetMaxIncomingConnections(&wrappers.Int64Value{Value: int64(maxConnections.(int))})
+						}
+						configMongoCfg.SetNet(&netMongoCfg)
+					}
+					if _, ok := d.GetOk("cluster_config.0.mongocfg.0.operation_profiling"); ok {
+						opProfilingMongoCfg := mongo_config.MongoCfgConfig6_0Enterprise_OperationProfiling{}
+						if mode, ok := d.GetOk("cluster_config.0.mongocfg.0.operation_profiling.0.mode"); ok {
+							modeInt := mongo_config.MongoCfgConfig6_0Enterprise_OperationProfiling_Mode_value[strings.ToUpper(mode.(string))]
+							opProfilingMongoCfg.SetMode(mongo_config.MongoCfgConfig6_0Enterprise_OperationProfiling_Mode(modeInt))
+						}
+
+						if opThreshold, ok := d.GetOk("cluster_config.0.mongocfg.0.operation_profiling.0.slow_op_threshold"); ok {
+							opProfilingMongoCfg.SetSlowOpThreshold(&wrappers.Int64Value{Value: int64(opThreshold.(int))})
+						}
+						configMongoCfg.SetOperationProfiling(&opProfilingMongoCfg)
+					}
+					if _, ok := d.GetOk("cluster_config.0.mongocfg.0.storage"); ok {
+						engineConfigMongoCfg := mongo_config.MongoCfgConfig6_0Enterprise_Storage_WiredTiger_EngineConfig{}
+						wiredTigerMongoCfg := mongo_config.MongoCfgConfig6_0Enterprise_Storage_WiredTiger{EngineConfig: &engineConfigMongoCfg}
+						storageMongoCfg := mongo_config.MongoCfgConfig6_0Enterprise_Storage{WiredTiger: &wiredTigerMongoCfg}
+
+						if cacheSize, ok := d.GetOk("cluster_config.0.mongocfg.0.storage.0.wired_tiger.0.cache_size_gb"); ok {
+							engineConfigMongoCfg.SetCacheSizeGb(&wrappers.DoubleValue{Value: float64(cacheSize.(float64))})
+						}
+						configMongoCfg.SetStorage(&storageMongoCfg)
+					}
+					hostTypes := getSetOfHostTypes(d)
+					var resourcesMongod, resourcesMongos, resourcesMongoCfg, resourcesMongoInfra *mongodb.Resources = getResources(d)
+					var mongod *mongodb.MongodbSpec6_0Enterprise_Mongod
+					var mongos *mongodb.MongodbSpec6_0Enterprise_Mongos
+					var mongocfg *mongodb.MongodbSpec6_0Enterprise_MongoCfg
+					var mongoinfra *mongodb.MongodbSpec6_0Enterprise_MongoInfra
+					mongod = &mongodb.MongodbSpec6_0Enterprise_Mongod{
+						Config:    &configMongod,
+						Resources: resourcesMongod,
+					}
+
+					if _, ok := hostTypes["MONGOS"]; ok {
+						mongos = &mongodb.MongodbSpec6_0Enterprise_Mongos{
+							Config:    &configMongos,
+							Resources: resourcesMongos,
+						}
+					}
+					if _, ok := hostTypes["MONGOCFG"]; ok {
+						mongocfg = &mongodb.MongodbSpec6_0Enterprise_MongoCfg{
+							Config:    &configMongoCfg,
+							Resources: resourcesMongoCfg,
+						}
+					}
+					if _, ok := hostTypes["MONGOINFRA"]; ok {
+						mongoinfra = &mongodb.MongodbSpec6_0Enterprise_MongoInfra{
+							ConfigMongocfg: &configMongoCfg,
+							ConfigMongos:   &configMongos,
+							Resources:      resourcesMongoInfra,
+						}
+					}
+					return &mongodb.ConfigSpec_MongodbSpec_6_0Enterprise{
+						MongodbSpec_6_0Enterprise: &mongodb.MongodbSpec6_0Enterprise{
+							Mongod:     mongod,
+							Mongos:     mongos,
+							Mongocfg:   mongocfg,
+							Mongoinfra: mongoinfra,
+						},
+					}
+				},
+			}
+		}
+	case "6.0":
+		{
+			return &MongodbSpecHelper{
+
+				FlattenResources: func(c *mongodb.ClusterConfig, d *schema.ResourceData) (map[string]interface{}, error) {
+					spec := c.Mongodb.(*mongodb.ClusterConfig_Mongodb_6_0).Mongodb_6_0
+					resources := map[string]interface{}{}
+					if _, ok := d.GetOk("resources"); ok {
+						if spec.Mongod != nil {
+							resources["resources"] = flattenMongoDBResources(spec.Mongod.Resources)
+							return resources, nil
+						}
+						if spec.Mongos != nil {
+							resources["resources"] = flattenMongoDBResources(spec.Mongos.Resources)
+							return resources, nil
+						}
+						if spec.Mongocfg != nil {
+							resources["resources"] = flattenMongoDBResources(spec.Mongocfg.Resources)
+							return resources, nil
+						}
+						if spec.Mongoinfra != nil {
+							resources["resources"] = flattenMongoDBResources(spec.Mongoinfra.Resources)
+							return resources, nil
+						}
+					} else {
+						if spec.Mongod != nil {
+							resources["resources_mongod"] = flattenMongoDBResources(spec.Mongod.Resources)
+						}
+						if spec.Mongos != nil {
+							resources["resources_mongos"] = flattenMongoDBResources(spec.Mongos.Resources)
+						}
+						if spec.Mongocfg != nil {
+							resources["resources_mongocfg"] = flattenMongoDBResources(spec.Mongocfg.Resources)
+						}
+						if spec.Mongoinfra != nil {
+							resources["resources_mongoinfra"] = flattenMongoDBResources(spec.Mongoinfra.Resources)
+						}
+					}
+					if len(resources) == 0 {
+						return nil, fmt.Errorf("Non empty service not found in mongo spec")
+					}
+					return resources, nil
+				},
+
+				FlattenMongod: func(c *mongodb.ClusterConfig, d *schema.ResourceData) ([]map[string]interface{}, error) {
+					mongod := c.Mongodb.(*mongodb.ClusterConfig_Mongodb_6_0).Mongodb_6_0.Mongod
+					if mongod != nil {
+						user_config := mongod.GetConfig().GetUserConfig()
+						result := map[string]interface{}{}
+
+						if net := user_config.GetNet(); net != nil {
+							flattenNet := map[string]interface{}{}
+							if maxIncomingConnections := net.GetMaxIncomingConnections(); maxIncomingConnections != nil {
+								flattenNet["max_incoming_connections"] = maxIncomingConnections.GetValue()
+							}
+							result["net"] = []map[string]interface{}{flattenNet}
+						}
+
+						if storage := user_config.GetStorage(); storage != nil {
+							flattenStorage := map[string]interface{}{}
+							if wiredTiger := storage.GetWiredTiger(); wiredTiger != nil {
+								flattenWiredTiger := map[string]interface{}{}
+								if engineConfig := wiredTiger.GetEngineConfig(); engineConfig != nil {
+									if cacheSize := engineConfig.GetCacheSizeGb(); cacheSize != nil {
+										flattenWiredTiger["cache_size_gb"] = cacheSize.GetValue()
+									}
+								}
+								if collectionConfig := wiredTiger.GetCollectionConfig(); collectionConfig != nil {
+									if blockCompressor := collectionConfig.GetBlockCompressor(); blockCompressor != 0 {
+										flattenWiredTiger["block_compressor"] = blockCompressor.String()
+									}
+								}
+								flattenStorage["wired_tiger"] = []map[string]interface{}{flattenWiredTiger}
+							}
+
+							if journal := storage.GetJournal(); journal != nil {
+								flattenJournal := map[string]interface{}{}
+								if commitInterval := journal.GetCommitInterval(); commitInterval != nil {
+									flattenJournal["commit_interval"] = commitInterval.GetValue()
+								}
+								flattenStorage["journal"] = []map[string]interface{}{flattenJournal}
+							}
+							result["storage"] = []map[string]interface{}{flattenStorage}
+						}
+
+						if opProfiling := user_config.GetOperationProfiling(); opProfiling != nil {
+							flattenOpProfiling := map[string]interface{}{}
+							if mode := opProfiling.GetMode(); mode != 0 {
+								flattenOpProfiling["mode"] = mode.String()
+							}
+							if opThreshold := opProfiling.GetSlowOpThreshold(); opThreshold != nil {
+								flattenOpProfiling["slow_op_threshold"] = opThreshold.GetValue()
+							}
+							result["operation_profiling"] = []map[string]interface{}{flattenOpProfiling}
+						}
+
+						return []map[string]interface{}{result}, nil
+					}
+					return []map[string]interface{}{}, nil
+				},
+
+				FlattenMongos: func(c *mongodb.ClusterConfig, d *schema.ResourceData) ([]map[string]interface{}, error) {
+					mongos := c.Mongodb.(*mongodb.ClusterConfig_Mongodb_6_0).Mongodb_6_0.Mongos
+					if mongos != nil {
+						userConfig := mongos.GetConfig().GetUserConfig()
+						result := map[string]interface{}{}
+
+						if net := userConfig.GetNet(); net != nil {
+							flattenNet := map[string]interface{}{}
+							if maxIncomingConnections := net.GetMaxIncomingConnections(); maxIncomingConnections != nil {
+								flattenNet["max_incoming_connections"] = maxIncomingConnections.GetValue()
+							}
+							result["net"] = []map[string]interface{}{flattenNet}
+						}
+						return []map[string]interface{}{result}, nil
+					}
+					return []map[string]interface{}{}, nil
+				},
+
+				FlattenMongocfg: func(c *mongodb.ClusterConfig, d *schema.ResourceData) ([]map[string]interface{}, error) {
+					mongocfg := c.Mongodb.(*mongodb.ClusterConfig_Mongodb_6_0).Mongodb_6_0.Mongocfg
+					if mongocfg != nil {
+						userConfig := mongocfg.GetConfig().GetUserConfig()
+						result := map[string]interface{}{}
+
+						if net := userConfig.GetNet(); net != nil {
+							flattenNet := map[string]interface{}{}
+							if maxIncomingConnections := net.GetMaxIncomingConnections(); maxIncomingConnections != nil {
+								flattenNet["max_incoming_connections"] = maxIncomingConnections.GetValue()
+							}
+							result["net"] = []map[string]interface{}{flattenNet}
+						}
+
+						if storage := userConfig.GetStorage(); storage != nil {
+							flattenStorage := map[string]interface{}{}
+							if wiredTiger := storage.GetWiredTiger(); wiredTiger != nil {
+								flattenWiredTiger := map[string]interface{}{}
+								if engineConfig := wiredTiger.GetEngineConfig(); engineConfig != nil {
+									if cacheSize := engineConfig.GetCacheSizeGb(); cacheSize != nil {
+										flattenWiredTiger["cache_size_gb"] = cacheSize.GetValue()
+									}
+								}
+								flattenStorage["wired_tiger"] = []map[string]interface{}{flattenWiredTiger}
+							}
+							result["storage"] = []map[string]interface{}{flattenStorage}
+						}
+
+						if opProfiling := userConfig.GetOperationProfiling(); opProfiling != nil {
+							flattenOpProfiling := map[string]interface{}{}
+							if mode := opProfiling.GetMode(); mode != 0 {
+								flattenOpProfiling["mode"] = mode.String()
+							}
+							if opThreshold := opProfiling.GetSlowOpThreshold(); opThreshold != nil {
+								flattenOpProfiling["slow_op_threshold"] = opThreshold.GetValue()
+							}
+							result["operation_profiling"] = []map[string]interface{}{flattenOpProfiling}
+						}
+
+						return []map[string]interface{}{result}, nil
+					}
+					return []map[string]interface{}{}, nil
+				},
+
+				Expand: func(d *schema.ResourceData) mongodb.ConfigSpec_MongodbSpec {
+					configMongod := mongo_config.MongodConfig6_0{}
+					configMongos := mongo_config.MongosConfig6_0{}
+					configMongoCfg := mongo_config.MongoCfgConfig6_0{}
+
+					if _, ok := d.GetOk("cluster_config.0.mongod.0.net"); ok {
+						netMongod := mongo_config.MongodConfig6_0_Network{}
+						if maxConnections, ok := d.GetOk("cluster_config.0.mongod.0.net.0.max_incoming_connections"); ok {
+							netMongod.SetMaxIncomingConnections(&wrappers.Int64Value{Value: int64(maxConnections.(int))})
+						}
+						configMongod.SetNet(&netMongod)
+					}
+					if _, ok := d.GetOk("cluster_config.0.mongod.0.operation_profiling"); ok {
+						opProfilingMongod := mongo_config.MongodConfig6_0_OperationProfiling{}
+
+						if mode, ok := d.GetOk("cluster_config.0.mongod.0.operation_profiling.0.mode"); ok {
+							modeInt := mongo_config.MongodConfig6_0_OperationProfiling_Mode_value[strings.ToUpper(mode.(string))]
+							opProfilingMongod.SetMode(mongo_config.MongodConfig6_0_OperationProfiling_Mode(modeInt))
+						}
+
+						if opThreshold, ok := d.GetOk("cluster_config.0.mongod.0.operation_profiling.0.slow_op_threshold"); ok {
+							opProfilingMongod.SetSlowOpThreshold(&wrappers.Int64Value{Value: int64(opThreshold.(int))})
+						}
+						configMongod.SetOperationProfiling(&opProfilingMongod)
+					}
+					if _, ok := d.GetOk("cluster_config.0.mongod.0.storage"); ok {
+						engineConfigMongod := mongo_config.MongodConfig6_0_Storage_WiredTiger_EngineConfig{}
+						collectionConfigMongod := mongo_config.MongodConfig6_0_Storage_WiredTiger_CollectionConfig{}
+						journalMongod := mongo_config.MongodConfig6_0_Storage_Journal{}
+						wiredTigerMongod := mongo_config.MongodConfig6_0_Storage_WiredTiger{
+							EngineConfig:     &engineConfigMongod,
+							CollectionConfig: &collectionConfigMongod,
+						}
+						storageMongod := mongo_config.MongodConfig6_0_Storage{
+							WiredTiger: &wiredTigerMongod,
+							Journal:    &journalMongod,
+						}
+						if cacheSize, ok := d.GetOk("cluster_config.0.mongod.0.storage.0.wired_tiger.0.cache_size_gb"); ok {
+							engineConfigMongod.SetCacheSizeGb(&wrappers.DoubleValue{Value: float64(cacheSize.(float64))})
+						}
+						if blockCompressor, ok := d.GetOk("cluster_config.0.mongod.0.storage.0.wired_tiger.0.block_compressor"); ok {
+							blockCompressorInt := mongo_config.MongodConfig6_0_Storage_WiredTiger_CollectionConfig_Compressor_value[strings.ToUpper(blockCompressor.(string))]
+							collectionConfigMongod.SetBlockCompressor(
+								mongo_config.MongodConfig6_0_Storage_WiredTiger_CollectionConfig_Compressor(blockCompressorInt),
+							)
+						}
+
+						if commitInterval, ok := d.GetOk("cluster_config.0.mongod.0.storage.0.journal.0.commit_interval"); ok {
+							journalMongod.SetCommitInterval(&wrappers.Int64Value{Value: int64(commitInterval.(int))})
+						}
+						configMongod.SetStorage(&storageMongod)
+					}
+					if _, ok := d.GetOk("cluster_config.0.mongos.0.net"); ok {
+
+						netMongos := mongo_config.MongosConfig6_0_Network{}
+						if maxConnections, ok := d.GetOk("cluster_config.0.mongos.0.net.0.max_incoming_connections"); ok {
+							netMongos.SetMaxIncomingConnections(&wrappers.Int64Value{Value: int64(maxConnections.(int))})
+						}
+						configMongos.SetNet(&netMongos)
+					}
+					if _, ok := d.GetOk("cluster_config.0.mongocfg.0.net"); ok {
+						netMongoCfg := mongo_config.MongoCfgConfig6_0_Network{}
+						if maxConnections, ok := d.GetOk("cluster_config.0.mongocfg.0.net.0.max_incoming_connections"); ok {
+							netMongoCfg.SetMaxIncomingConnections(&wrappers.Int64Value{Value: int64(maxConnections.(int))})
+						}
+						configMongoCfg.SetNet(&netMongoCfg)
+					}
+					if _, ok := d.GetOk("cluster_config.0.mongocfg.0.operation_profiling"); ok {
+						opProfilingMongoCfg := mongo_config.MongoCfgConfig6_0_OperationProfiling{}
+						if mode, ok := d.GetOk("cluster_config.0.mongocfg.0.operation_profiling.0.mode"); ok {
+							modeInt := mongo_config.MongoCfgConfig6_0_OperationProfiling_Mode_value[strings.ToUpper(mode.(string))]
+							opProfilingMongoCfg.SetMode(mongo_config.MongoCfgConfig6_0_OperationProfiling_Mode(modeInt))
+						}
+
+						if opThreshold, ok := d.GetOk("cluster_config.0.mongocfg.0.operation_profiling.0.slow_op_threshold"); ok {
+							opProfilingMongoCfg.SetSlowOpThreshold(&wrappers.Int64Value{Value: int64(opThreshold.(int))})
+						}
+						configMongoCfg.SetOperationProfiling(&opProfilingMongoCfg)
+					}
+					if _, ok := d.GetOk("cluster_config.0.mongocfg.0.storage"); ok {
+						engineConfigMongoCfg := mongo_config.MongoCfgConfig6_0_Storage_WiredTiger_EngineConfig{}
+						wiredTigerMongoCfg := mongo_config.MongoCfgConfig6_0_Storage_WiredTiger{EngineConfig: &engineConfigMongoCfg}
+						storageMongoCfg := mongo_config.MongoCfgConfig6_0_Storage{WiredTiger: &wiredTigerMongoCfg}
+
+						if cacheSize, ok := d.GetOk("cluster_config.0.mongocfg.0.storage.0.wired_tiger.0.cache_size_gb"); ok {
+							engineConfigMongoCfg.SetCacheSizeGb(&wrappers.DoubleValue{Value: float64(cacheSize.(float64))})
+						}
+						configMongoCfg.SetStorage(&storageMongoCfg)
+					}
+					hostTypes := getSetOfHostTypes(d)
+					var resourcesMongod, resourcesMongos, resourcesMongoCfg, resourcesMongoInfra *mongodb.Resources = getResources(d)
+					var mongod *mongodb.MongodbSpec6_0_Mongod
+					var mongos *mongodb.MongodbSpec6_0_Mongos
+					var mongocfg *mongodb.MongodbSpec6_0_MongoCfg
+					var mongoinfra *mongodb.MongodbSpec6_0_MongoInfra
+					mongod = &mongodb.MongodbSpec6_0_Mongod{
+						Config:    &configMongod,
+						Resources: resourcesMongod,
+					}
+
+					if _, ok := hostTypes["MONGOS"]; ok {
+						mongos = &mongodb.MongodbSpec6_0_Mongos{
+							Config:    &configMongos,
+							Resources: resourcesMongos,
+						}
+					}
+					if _, ok := hostTypes["MONGOCFG"]; ok {
+						mongocfg = &mongodb.MongodbSpec6_0_MongoCfg{
+							Config:    &configMongoCfg,
+							Resources: resourcesMongoCfg,
+						}
+					}
+					if _, ok := hostTypes["MONGOINFRA"]; ok {
+						mongoinfra = &mongodb.MongodbSpec6_0_MongoInfra{
+							ConfigMongocfg: &configMongoCfg,
+							ConfigMongos:   &configMongos,
+							Resources:      resourcesMongoInfra,
+						}
+					}
+
+					return &mongodb.ConfigSpec_MongodbSpec_6_0{
+						MongodbSpec_6_0: &mongodb.MongodbSpec6_0{
+							Mongod:     mongod,
+							Mongos:     mongos,
+							Mongocfg:   mongocfg,
+							Mongoinfra: mongoinfra,
+						},
+					}
+				},
+			}
+		}
 	case "5.0-enterprise":
 		{
 			return &MongodbSpecHelper{